feat: added customizable default value for jsonMapper
