+++ packages/xlsx-import/src/mappers/jsonMapper.ts
@@ -1,9 +1,33 @@
-type JsonMapper = <TJsonResult>(value: string) => TJsonResult | null;
 
-export const jsonMapper: JsonMapper = <TJsonResult extends any>(value: string): TJsonResult | null => {
-    try {
-        return JSON.parse(value);
-    } catch (e) {
-        return null;
+type callSignature = <TJsonResult, TDefaultResult>(value: string) => TJsonResult | TDefaultResult;
+
+type JsonMapper = {
+    default: <TDefaultResult>(defaultResult: TDefaultResult) => JsonMapper
+} & callSignature;
+
+interface IJsonMapperOptions {
+    default: any;
+}
+
+const factory = (options: Readonly<IJsonMapperOptions>): JsonMapper => {
+    const mapper = (json: string) => {
+        try {
+            return JSON.parse(json);
+        } catch (e) {
+            return options.default
+        }
     }
+
+    mapper.default = <TDefaultResult>(defaultResult: TDefaultResult) => factory({
+        ...options,
+        default: defaultResult
+    });
+
+    return mapper;
 };
+
+
+
+export const jsonMapper: JsonMapper = factory({
+    default: null,
+});
